Rename network

Approved-by: Bradley Wootton
Approved-by: Leo Denham
Approved-by: Ethan Katte

+++ src/backend/net_tools.py
@@ -19,7 +19,7 @@ from job import Job
 from MAC_table import MAC_table
 from device import Device
 from platform import system
-import nmap, socket, netifaces, threading, sys, signal, time
+import nmap, socket, netifaces, threading, sys, signal, time, subprocess, os
 
 if system() != "Darwin":
     import pywifi
@@ -67,16 +67,22 @@ class Net_tools:
     # --------------------------------------------- ON START -------------------------------------- #
 
     # Prepares the backend to run scans on the current network
-    def new_network(self, name):
+    def new_network(self):
 
-        self.name = name
+        ssid = self.get_ssid()
+
+        if not ssid:
+            return False
+
+        self.name = ssid
 
         # Deleted network if it already exists
         if self.db.contains_network(self.name):
             self.db.delete_network(self.name)
 
         # Creates a new table in the database for the current network if it doesnt already exist
-        self.db.register_network(self.gateway_mac, self.get_ssid(), self.name)
+        if not self.db.register_network(self.gateway_mac, ssid, self.name):
+            return False
 
         self.client_mac = Ether().src
 
@@ -97,6 +103,8 @@ class Net_tools:
         DNS_sniffer.daemon = True
         DNS_sniffer.start()
 
+        return True
+
 
     # Signal handler to gracefully end the threadpool on shutdown
     def cleanup(
@@ -112,9 +120,9 @@ class Net_tools:
     def get_ssid(self):
 
         current_system = system()
+
         # MacOS - Get by running the airport program.
         if current_system == "Darwin":
-            import subprocess
 
             process = subprocess.Popen(
                 [
@@ -129,32 +137,16 @@ class Net_tools:
                 if " SSID" in line:
                     return line.split(": ")[1]
 
-        wifi = pywifi.PyWiFi()
-        ifaces = wifi.interfaces()
+        elif current_system == "Linux":
 
-        if len(ifaces) < 1:
-            return "Disconnected"
+            return os.popen("iwconfig " + conf.iface + " | grep ESSID | awk '{print $4}' | sed 's/" + '"' + "//g' | sed 's/.*ESSID://'").read()[:-1]
 
-        iface = ifaces[0]
-        # Exponential dropoff
-        results = []
-        for exp in range(6):
-            # Start scan and wait for results
-            iface.scan()
-            time.sleep(0.1 * pow(2, exp))
-            results = iface.scan_results()
+        elif current_system == "Windows":
 
-            if len(results) > 0:
-                break
+            out = os.popen('netsh wlan show interfaces | findstr /c:" SSID"').read()[:-1]
+            return out.split(":")[-1][1:]
 
-        if results:
-            # TODO fix
-            # print(
-            #     f"Seems to return the strongest signal, not the current connected one... {[x.ssid for x in results]}"
-            # )
-            return results[0].ssid
-
-        return "Disconnected"
+        return "OS UNSUPPORTED"
 
     # ---------------------------------------------- MAC VENDOR ---------------------------------------------- #
 
@@ -505,20 +497,10 @@ class Net_tools:
         counter_ptr = [0]
         returns = [-1] * len(devices.keys())
 
-        # Counts the number of dispatched jobs
-        dispatched = len(devices.keys())
-
         # The current job, for referencing the return location for the thread
         job_counter = 0
 
         for device in devices.values():
-            # If hostname has been found in wlan DNSRR sniffer, use it instead
-            if device.hostname != "unknown":
-                self.lb.increment()
-                dispatched -= 1
-                returns[job_counter] = device.hostname
-                job_counter += 1
-                continue
 
             # Create job and add to threadpool queue for execution
             job = Job(
@@ -538,9 +520,9 @@ class Net_tools:
 
         # Wait for all jobs to be comleted
         mutex.acquire()
-        while counter_ptr[0] < dispatched:
+        while counter_ptr[0] < len(devices.keys()):
             cond.wait()
-            self.lb.set_progress(counter_ptr[0] + (len(devices.keys()) - dispatched))
+            self.lb.set_progress(counter_ptr[0])
             self.lb.show()
 
         mutex.release()
@@ -548,7 +530,7 @@ class Net_tools:
         # Add returned hostnames to devices, save to database
         job_counter = 0
         for device in devices.values():
-            if device.hostname != "unkown":
+            if returns[job_counter] != "unkown":
                 device.hostname = returns[job_counter]
                 self.db.save_device(self.name, device)
             job_counter += 1
@@ -622,8 +604,9 @@ class Net_tools:
                         print("[DEBUG] err in wlan sniff")
                         return
 
-                    device.hostname = name
-                    self.db.save_device(self.name, device)
+                    if device.hostname == "unknown":
+                        device.hostname = name
+                        self.db.save_device(self.name, device)
 
 
     def run_wlan_sniffer(self, iface):
